+++ HPXMLtoOpenStudio/resources/constants.rb
@@ -356,6 +356,10 @@ class Constants
     return 'room ac'
   end
 
+  def self.ObjectNameShowers
+    return 'hot_water_showers'
+  end
+
   def self.ObjectNameSolarHotWater
     return 'solar hot water'
   end
@@ -370,6 +374,10 @@ class Constants
 
   def self.ObjectNameUnmetHoursProgram
     return 'unmet hours program'
+  end
+
+  def self.ObjectNameUnmetLoadsProgram
+    return 'unmet loads program'
   end
 
   def self.ObjectNameWaterHeater