+++ HPXMLtoOpenStudio/tests/test_schedules.rb
@@ -51,6 +51,7 @@ class HPXMLtoOpenStudioSchedulesTest < Minitest::Test
     args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base.xml'))
     model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
 
+
     schedule_constants = 12
     schedule_rulesets = 17
     schedule_fixed_intervals = 1
@@ -191,7 +192,7 @@ class HPXMLtoOpenStudioSchedulesTest < Minitest::Test
     model.getScheduleFiles.each do |schedule_file|
       schedule_file_names << "#{schedule_file.name}"
     end
-    assert_equal(11, schedule_file_names.size)
+    assert_equal(12, schedule_file_names.size)
 
     schedules_paths = hpxml_bldg.header.schedules_filepaths.collect { |sfp|
       FilePath.check_path(sfp,